Correctly update `:dirty` for values that are collections
`(clojure.data/diff values initial-values` returns a tuple `[only-in-values, only-in-initial-values, in-both]`. If one value of the map `values` has a collection (set or vector for example), and the user removes a value in this collection, the `diff` call will show `nil` as the first tuple element.

By merging both the first and second element of the `diff` result `dirty` will still work as before for simple (non-collection) values and also for collection values.